Move `converter` module into `model` package
