--- bankroll/converter/__init__.py
@@ -1,3 +1,0 @@
-from .converter import dataframeForModelObjects
-
-__all__ = ["dataframeForModelObjects"]
